fetch semgrep app token better
